requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.0
executing==2.2.0
fastapi==0.115.11
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
groq==0.19.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.3.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-groq==0.2.5
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
langgraph==0.3.11
langgraph-checkpoint==2.0.20
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
motor==3.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
orjson==3.10.15
packaging==24.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.2
pypdf==5.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wheel==0.45.1
yarl==1.18.3
zstandard==0.23.0
